Move ParseImageReferences to cri util
Avoids importing image service for utility function.
